backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
slowapi==0.1.9
python-multipart==0.0.12
anthropic==0.40.0
